Refactor: Move retrieve_graph and construct_code_graph_context functions from localize.py and repair.py to get_repo_structure.py for code reuse
